Creating CSS Roll Over Menu Difficulties.
I have looked over my code many times, and I can't find the mistake I am making. The feature that is NOT working is when you hover over the Portfolio tab a drop down menu is supposed to appear. But it doesn't on mine and I am not sure why ??

Any clues?